refactor(geolens_pkg): rename utils.py → optim_init.py; move init_constraints to post_computation
Rename geolens_pkg/utils.py to optim_init.py so the three-file
optimization workflow is discoverable as a group: optim_init (build
starting point) → optim (gradient loop) → optim_ops (discrete edits).
Also resolves naming collision with optics/utils.py.

Move init_constraints() call from get_optimizer() into
post_computation() so constraints are always refreshed when lens
geometry is loaded or modified, not only when an optimizer is created.
