+++ deeplens/optics/geolens_pkg/__init__.py
@@ -1,4 +1,4 @@
-from .utils import create_lens
+from .optim_init import create_lens
 
 __all__ = [
     "create_lens",
